WIP :Lookup questions outside of summary structure build
