+++ app/views/contexts/summary/block.py
@@ -5,9 +5,11 @@ from app.views.contexts.summary.question import Question
 class Block:
     @staticmethod
     def get_question(
-        block_schema, answer_store, list_store, metadata, schema, current_location=None
+        block_id, answer_store, list_store, metadata, schema, current_location=None
     ):
         """ Taking question variants into account, return the question which was displayed to the user """
+        block_schema = schema.get_block(block_id)
+
         for variant in block_schema.get('question_variants', []):
             display_variant = evaluate_when_rules(
                 variant.get('when'),
